bot-requirements.txt
python-telegram-bot==20.7
python-telegram-bot[webhooks]==20.7 